+++ src/tests/matchers/index.ts
@@ -1,2 +1,2 @@
 export * from './date-to-equal-within-interval';
-export * from './focused-day-to-equal';
+export * from './focused-date-to-equal';