+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './dates';
+export * from './find-last-index';
 export * from './is-instance-of';
 export * from './pad-start';
 export * from './parse-picker-input';