+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - `/jest.config.ts`
 - Add unit tests for base `Picker` class
 - Add `Picker.thenUpdateItems()` method to group state changes before updating items
+- Add `findLastIndex()` utility function
 
 ### Changed
 - Change `DatePickle.constructor` signature